+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changes in version 0.7 - 2020-??-??
 
  - Add `onionperf measure --drop-guards` parameter to use and drop
-   guards after a given number of hours. Implements #33399.
+   guards and circuit build timeouts every given number of hours, if
+   supported by the Tor version. Implements #33399.
  - Remove the `onionperf measure --oneshot` switch and replace it with
    new switches `--tgen-pause-initial`, `--tgen-pause-between`,
    `--tgen-transfer-size`, and `--tgen-num-transfers ` to further